feat: Improve Drizzle Studio connection to local Cloudflare D1 database
This commit adds improvements to the connection between Drizzle Studio and the local Cloudflare D1 database. Previously, it was necessary to manually specify the SQLite path in Wrangler to use the local database from Drizzle Studio. With this update, a new npm script `db:studio` has been added, which uses the `cross-env` command to search for the SQLite database file in the Wrangler folder and register it as the `LOCAL_DB_PATH` environment variable. Additionally, a new npm script `db:studio:prod` has been added to connect to the remote D1 database. The `drizzle.config.ts` file has also been updated to use the `better-sqlite3` driver when the `LOCAL_DB_PATH` environment variable exists, allowing Drizzle Studio to connect to the local database. Otherwise, the `d1` driver is used as usual.